Support for VMs merge with remapper + new sdk
Tyche now supports enclaves and VMs thanks to the new kvm-backend
themis.

It also offers a new sdk with two alternative backends:
1. skd tyche: directly interacts with the tyche driver to create
   domains. DOES NOT SUPPORT INTERRUPT ROUTING.
2. sdk kvm: leverages kvm to create tds. SUPPORTS interrupts
   transparently.

I am gonna clean a lot of things, there are too many back and forth with
adding value and resetting afterwards etc.
Let's just clean that all up.

2. Changing the runtime sdk

We do not need the frame_t anymore, we look for it directly by talking
to tyche. Why? The capa index can change if we had interrupts and were
preempted. This is also a proof of concept of integrating the capa lib
inside and enclave and enumerating its own resource which is a first
step towards checking your are in a confidential environment or spawning
a "nested" domain.

+++ C/libraries/pts/src/lib.c
@@ -1,6 +1,7 @@
 #include "pts_api.h"
 
 #include "common.h"
+#include "common_log.h"
 
 /// From a VA addr and a level, get the current index.
 index_t get_index(addr_t addr, level_t level, pt_profile_t* profile)
@@ -24,7 +25,7 @@ int pt_walk_page_range(entry_t root, level_t level, addr_t start, addr_t end, pt
   curr_va = start;
   s = get_index(start, level, profile);
   va_root = (entry_t*) profile->pa_to_va(root, profile);
-  DEBUG("va_root: %llx", va_root);
+  DEBUG("va_root: %llx", (usize) va_root);
   for (i = s; i < profile->nb_entries; i++) {
     // Compute the current virtual address.
     curr_va = (start & ~(profile->masks[level]))
